some refactor code
